# docs/api/openapi.yaml
components:
  schemas:
...
    ApiMeta:
      additionalProperties: true
      description: >-
        List metadata (docs/07_API.md §19). The server normalizes page/limit (max server-enforced limit: 100) and
        computes total/totalPages/hasNext. Clients must never derive totals themselves.
      properties:
        hasNext:
          type: boolean
        limit:
          maximum: 100
          minimum: 1
          type: integer
        page:
          minimum: 1
          type: integer
        total:
          minimum: 0
          type: integer
        totalPages:
          minimum: 0
          type: integer
      type: object
...
    ErrorCode:
      enum:
        - AUTH_REQUIRED
        - AUTH_INVALID
        - FORBIDDEN
        - NOT_FOUND
        - VALIDATION_ERROR
        - CONFLICT
        - RATE_LIMITED
        - SUBSCRIPTION_REQUIRED
        - ENTITLEMENT_REQUIRED
        - INVALID_STATE_TRANSITION
        - UPLOAD_REJECTED
        - INTERNAL_ERROR
      type: string
    ErrorResponse:
      description: Canonical error envelope (docs/07_API.md §3).
      properties:
        error:
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            fields:
              additionalProperties:
                type: string
              description: Field-level validation details when applicable.
              type: object
            message:
              type: string
          required:
            - code
            - message
          type: object
      required:
        - error
      type: object
...
    ReviewSummaryDto:
      additionalProperties: false
      description: >-
        Public review for technician discovery. Only documented public fields: no customer identifiers, no moderation
        state.
      properties:
        comment:
          nullable: true
          type: string
        createdAt:
          format: date-time
          type: string
        id:
          format: uuid
          type: string
        rating:
          maximum: 5
          minimum: 1
          type: integer
        tags:
          items:
            type: string
          type: array
      required:
        - id
        - rating
        - comment
        - tags
        - createdAt
      type: object
...
    ServiceDto:
      additionalProperties: false
      description: Active service/specialty catalog entry.
      properties:
        applianceCategoryId:
          format: uuid
          type: string
        descriptionAr:
          nullable: true
          type: string
        id:
          format: uuid
          type: string
        nameAr:
          type: string
        slug:
          type: string
      required:
        - id
        - applianceCategoryId
        - nameAr
        - slug
        - descriptionAr
      type: object
    ServiceRequestDto:
      additionalProperties: false
      description: >-
        Role-scoped request detail. The request location (job information) is included for the customer owner and the
        targeted/assigned technician. Status history is bounded and append-only.
      properties:
        acceptedAt:
          format: date-time
          nullable: true
          type: string
        applianceCategoryId:
          format: uuid
          type: string
        cancelledAt:
          format: date-time
          nullable: true
          type: string
        completedAt:
          format: date-time
          nullable: true
          type: string
        createdAt:
          format: date-time
          type: string
        faultId:
          format: uuid
          nullable: true
          type: string
        history:
          items:
            $ref: '#/components/schemas/ServiceRequestStatusHistoryDto'
          type: array
        id:
          format: uuid
          type: string
        location:
          additionalProperties: false
          properties:
            addressText:
              nullable: true
              type: string
            city:
              nullable: true
              type: string
            label:
              nullable: true
              type: string
            latitude:
              type: number
            longitude:
              type: number
            region:
              nullable: true
              type: string
          required:
            - label
            - addressText
            - city
            - region
            - latitude
            - longitude
          type: object
        problemDescription:
          type: string
        problemTitle:
          nullable: true
          type: string
        scheduledAt:
          format: date-time
          nullable: true
          type: string
        serviceId:
          format: uuid
          nullable: true
          type: string
        startedAt:
          format: date-time
          nullable: true
          type: string
        status:
          $ref: '#/components/schemas/ServiceRequestStatus'
        technicianId:
          format: uuid
          nullable: true
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - status
        - problemTitle
        - problemDescription
        - applianceCategoryId
        - serviceId
        - faultId
        - technicianId
        - location
        - scheduledAt
        - acceptedAt
        - startedAt
        - completedAt
        - cancelledAt
        - createdAt
        - updatedAt
        - history
      type: object
    ServiceRequestStatus:
      enum:
        - pending
        - accepted
        - on_the_way
        - in_progress
        - completed
        - cancelled
      type: string
    ServiceRequestStatusHistoryDto:
      additionalProperties: false
      description: Append-only status transition record (docs/07_API.md §13).
      properties:
        changedByUserId:
          format: uuid
          nullable: true
          type: string
        createdAt:
          format: date-time
          type: string
        fromStatus:
          $ref: '#/components/schemas/ServiceRequestStatus'
          nullable: true
        id:
          format: uuid
          type: string
        toStatus:
          $ref: '#/components/schemas/ServiceRequestStatus'
      required:
        - id
        - fromStatus
        - toStatus
        - changedByUserId
        - createdAt
      type: object
    ServiceRequestSummaryDto:
      additionalProperties: false
      description: Role-scoped service-request list item.
      properties:
        applianceCategoryId:
          format: uuid
          type: string
        createdAt:
          format: date-time
          type: string
        faultId:
          format: uuid
          nullable: true
          type: string
        id:
          format: uuid
          type: string
        problemDescription:
          type: string
        problemTitle:
          nullable: true
          type: string
        scheduledAt:
          format: date-time
          nullable: true
          type: string
        serviceId:
          format: uuid
          nullable: true
          type: string
        status:
          $ref: '#/components/schemas/ServiceRequestStatus'
        technicianId:
          format: uuid
          nullable: true
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - status
        - problemTitle
        - problemDescription
        - applianceCategoryId
        - serviceId
        - faultId
        - technicianId
        - scheduledAt
        - createdAt
        - updatedAt
      type: object
...
    TechnicianAvailabilityStatus:
      enum:
        - available
        - busy
        - unavailable
      type: string
    TechnicianPublicDto:
      additionalProperties: false
      description: >-
        Public technician profile for discovery/detail. Only verified technicians are publicly visible. Excludes
        authentication data, private contact data, and internal moderation state.
      properties:
        availabilityStatus:
          $ref: '#/components/schemas/TechnicianAvailabilityStatus'
        avatarUrl:
          nullable: true
          type: string
        bio:
          nullable: true
          type: string
        completedServicesCount:
          type: integer
        displayName:
          nullable: true
          type: string
        experienceYears:
          type: integer
        id:
          format: uuid
          type: string
        ratingAverage:
          nullable: true
          type: number
        ratingCount:
          type: integer
        services:
          items:
            $ref: '#/components/schemas/TechnicianServiceDto'
          type: array
        verificationStatus:
          $ref: '#/components/schemas/VerificationStatus'
      required:
        - id
        - displayName
        - bio
        - avatarUrl
        - verificationStatus
        - availabilityStatus
        - experienceYears
        - completedServicesCount
        - ratingAverage
        - ratingCount
        - services
      type: object
    TechnicianServiceDto:
      additionalProperties: false
      description: A service/specialty offered by a technician (active only).
      properties:
        service:
          $ref: '#/components/schemas/ServiceDto'
      required:
        - service
      type: object
...
    VerificationStatus:
      enum:
        - pending
        - verified
        - rejected
        - suspended
      type: string
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: >-
    Al-Khabir backend contract. Generated from the NestJS controllers (ADR-0003). The canonical success envelope is `{
    data, meta? }`; errors use `{ error: { code, message, fields? } }` with the codes documented in docs/07_API.md §21.
  title: Al-Khabir API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /api/v1/service-requests:
    get:
      operationId: ServiceRequestsController_list
      parameters:
        - in: query
          name: page
          required: false
          schema:
            minimum: 1
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            maximum: 100
            minimum: 1
            type: integer
        - in: query
          name: status
          required: false
          schema:
            enum:
              - pending
              - accepted
              - on_the_way
              - in_progress
              - completed
              - cancelled
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceRequestSummaryDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Role-scoped request list (paginated).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing, malformed, or expired access token / role mismatch.
      security:
        - bearer: []
      tags:
        - service-requests
    post:
      operationId: ServiceRequestsController_create
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                appliance_category_id:
                  type: string
                fault_id:
                  type: string
                location_id:
                  type: string
                problem_description:
                  maxLength: 5000
                  minLength: 1
                  type: string
                problem_title:
                  maxLength: 255
                  minLength: 1
                  type: string
                scheduled_at: {}
                service_id:
                  type: string
                technician_id:
                  type: string
              required:
                - technician_id
                - appliance_category_id
                - problem_description
                - location_id
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceRequestDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Request created with status=pending (history recorded).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing, malformed, or expired access token / role mismatch.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Referenced technician/location/category/service/fault not found or not owned.
      security:
        - bearer: []
      tags:
        - service-requests
...
  /api/v1/service-requests/{id}/review:
    post:
      operationId: ServiceRequestReviewsController_create
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                comment:
                  maxLength: 2000
                  type: string
                rating:
                  maximum: 5
                  minimum: 1
                  type: integer
                tag_ids:
                  items:
                    type: string
                  type: array
              required:
                - rating
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ReviewSummaryDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Review created (owner, completed request, one per request).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing/malformed token or non-customer role.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request not found or not owned by this customer.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Review already exists / request not completed (canonical envelope).
      security:
        - bearer: []
      tags:
        - reviews
...
  /api/v1/technicians:
    get:
      operationId: TechniciansController_list
      parameters:
        - in: query
          name: page
          required: false
          schema:
            minimum: 1
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            maximum: 100
            minimum: 1
            type: integer
        - in: query
          name: q
          required: false
          schema:
            maxLength: 100
            minLength: 1
            type: string
        - in: query
          name: appliance_category_id
          required: false
          schema:
            type: string
        - in: query
          name: service_id
          required: false
          schema:
            type: string
        - in: query
          name: fault_id
          required: false
          schema:
            type: string
        - in: query
          name: rating_min
          required: false
          schema:
            maximum: 5
            minimum: 1
            type: integer
        - in: query
          name: availability
          required: false
          schema:
            enum:
              - available
              - busy
              - unavailable
            type: string
        - in: query
          name: sort
          required: false
          schema: {}
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/TechnicianPublicDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Verified technicians (paginated discovery list).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
      tags:
        - catalog
...
  /api/v1/technicians/{id}/reviews:
    get:
      operationId: TechnicianReviewsController_listForTechnician
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            minimum: 1
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            maximum: 100
            minimum: 1
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ReviewSummaryDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Public reviews for a verified technician (paginated).
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Technician not found or not publicly visible.
      tags:
        - reviews